# my-api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/books/{id}:
    get:
      tags:
      - book-rest-controller
      operationId: getBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Book"
    put:
      tags:
      - book-rest-controller
      operationId: updateBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Book"
    delete:
      tags:
      - book-rest-controller
      operationId: deleteBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
    patch:
      tags:
      - book-rest-controller
      operationId: partialUpdateBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Book"
  /api/books:
    get:
      tags:
      - book-rest-controller
      operationId: getAllBooks
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      tags:
      - book-rest-controller
      operationId: addBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Book"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Book"
components:
  schemas:
    Author:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Book:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        image_url:
          type: string
        publicationYear:
          type: integer
          format: int32
        publishers:
          type: array
          items:
            $ref: "#/components/schemas/Publisher"
        authors:
          type: array
          items:
            $ref: "#/components/schemas/Author"
        genre:
          type: string
          enum:
          - ART
          - BIOGRAPHY
          - CHILDREN
          - FICTION
          - HISTORY
          - MYSTERY
          - PHILOSOPHY
          - RELIGION
          - ROMANCE
          - SELFT_HELP
          - TECHNICAL
        amazonRating:
          type: number
          format: double
    Publisher:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string